requirements.txt
# Core dependencies
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0

# Data handling
xarray>=2023.0.0
netCDF4>=1.6.0
pyarrow>=14.0.0
h5netcdf>=1.2.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
redis>=5.0.0

# Vector database
faiss-cpu>=1.7.4
chromadb>=0.4.0
sentence-transformers>=2.2.0

# LLM & AI
langchain>=0.1.0
langchain-community>=0.0.10
ollama>=0.1.0

# Visualization
plotly>=5.18.0
folium>=0.15.0
streamlit-folium>=0.15.0
pydeck>=0.8.0
matplotlib>=3.8.0
seaborn>=0.13.0

# API
fastapi>=0.109.0
uvicorn>=0.25.0
httpx>=0.26.0

# Voice & Speech
openai-whisper>=20231117
sounddevice>=0.4.6
soundfile>=0.12.1

# Anomaly Detection
scikit-learn>=1.3.0
scipy>=1.11.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
requests>=2.31.0
aiofiles>=23.2.0
python-multipart>=0.0.6

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0

# Development
black>=23.12.0
isort>=5.13.0
